libraries/cloudharness-common/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
idna==2.8
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.6.8
requests==2.21.0
ruamel-yaml==0.15.46
six==1.12.0
urllib3==1.24.1
pykafka==2.8.0
pyaml
kafka-python
kubernetes